rename package to lightweight_lidar_only_simulator
