+++ include/lightweight_lidar_only_simulator/ackermann_kinematics.hpp
@@ -1,8 +1,8 @@
 #pragma once
 
-#include "racecar_simulator/pose_2d.hpp"
+#include "lightweight_lidar_only_simulator/pose_2d.hpp"
 
-namespace racecar_simulator {
+namespace lightweight_lidar_only_simulator {
 
 class AckermannKinematics {
