+++ include/lightweight_lidar_only_simulator/scan_simulator_2d.hpp
@@ -2,9 +2,9 @@
 
 #include <random>
 
-#include "racecar_simulator/pose_2d.hpp"
+#include "lightweight_lidar_only_simulator/pose_2d.hpp"
 
-namespace racecar_simulator {
+namespace lightweight_lidar_only_simulator {
 
 class ScanSimulator2D {
