+++ launch/distance_transform_visualizer.launch.py
@@ -4,10 +4,10 @@ from ament_index_python.packages import get_package_share_directory
 
 
 def generate_launch_description():
-    package = get_package_share_directory('racecar_simulator')
+    package = get_package_share_directory('lightweight_lidar_only_simulator')
     ld = launch.LaunchDescription([
         launch_ros.actions.Node(
-            package='racecar_simulator',
+            package='lightweight_lidar_only_simulator',
             executable='distance_transform_visualizer',
             name='distance_transform_visualizer',
             output='screen',
