+++ launch/racecar_model.launch.py
@@ -6,7 +6,7 @@ from launch_ros.substitutions import FindPackageShare
 from ament_index_python.packages import get_package_share_directory
 
 def generate_launch_description():
-    package = get_package_share_directory('racecar_simulator')
+    package = get_package_share_directory('lightweight_lidar_only_simulator')
     robot_description_content = Command([
         PathJoinSubstitution([FindExecutable(name="xacro")]),
         " ",
