+++ package.xml
@@ -1,9 +1,9 @@
 <package format="2">
-  <name>racecar_simulator</name>
+  <name>lightweight_lidar_only_simulator</name>
   <version>0.0.0</version>
-  <description>A simulator of the MIT racecar.</description>
+  <description>A lightweight simulator for a car with lidar</description>
 
-  <maintainer email="tfh@mit.edu">tfh</maintainer>
+  <maintainer email="imperial.driverless@gmail.com">Imperial Driverless</maintainer>
 
   <license>MIT</license>
   <buildtool_depend>ament_cmake</buildtool_depend>
